requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
mysqlclient>=2.2 ; platform_system != 'Windows'
PyMySQL>=1.1
python-dotenv>=1.0
cryptography>=43.0
requests>=2.32
xhtml2pdf>=0.2.15
